Migrate plan_tool_followup family (4 methods) to actor_loop_flow
Move the plan_tool_followup phase dispatcher and 3 private sub-helpers
out of impl Agent into actor_loop_flow.rs:

- handle_actor_loop_plan_tool_followup (pub(super) dispatcher)
- handle_plan_file_edit_followup (private helper)
- handle_plan_exploration_followup (private helper)
- handle_actor_loop_plan_exploration_only_turn (private helper)

The 3 sub-helpers are module-private (no pub(super)) since their
only consumer is the dispatcher.

Visibility upgrade on turn.rs free fn:

- pub(super) on log_plan_stall (used by both exploration_only_turn
  and the still-in-turn.rs handle_actor_loop_completion)

Drop now-unused production imports of handle_non_progress_plan_edit_fallback
and plan_tool_followup_done_message; the former loses all callers (the
plan-file-edit followup was its sole production consumer) while the
latter is still used in the in-file regression test and is moved under
#[cfg(test)].

Part of #680 / #681 incremental method migration.
